Refactor: Protostar -> IVC
