+++ symbolic-base/src/ZkFold/Base/Protocol/IVC.hs
@@ -1,0 +1,3 @@
+module ZkFold.Base.Protocol.IVC (module Protostar) where
+
+import           ZkFold.Base.Protocol.IVC.Internal as Protostar
