+++ symbolic-base/src/ZkFold/Base/Protocol/IVC/RecursiveCircuit.hs
@@ -1,28 +1,27 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE TypeOperators       #-}
 
-module ZkFold.Base.Protocol.Protostar.RecursiveCircuit where
+module ZkFold.Base.Protocol.IVC.RecursiveCircuit where
 
-import           GHC.Generics                                          (Par1 (..), U1 (..), type (:.:) (..), (:*:) (..))
-import           Prelude                                               hiding (Num (..), drop, head, replicate, take,
-                                                                        zipWith)
+import           GHC.Generics                                    (Par1 (..), U1 (..), type (:.:) (..), (:*:) (..))
+import           Prelude                                         hiding (Num (..), drop, head, replicate, take, zipWith)
 
 import           ZkFold.Base.Algebra.Basic.Class
-import           ZkFold.Base.Algebra.Basic.Number                      (KnownNat, type (+), type (-), value)
-import           ZkFold.Base.Data.ByteString                           (Binary)
-import           ZkFold.Base.Data.Vector                               (Vector, unsafeToVector)
-import           ZkFold.Base.Protocol.Protostar.ArithmetizableFunction (ArithmetizableFunction (ArithmetizableFunction))
-import           ZkFold.Base.Protocol.Protostar.Commit                 (HomomorphicCommit)
-import           ZkFold.Base.Protocol.Protostar.CommitOpen             (CommitOpen (..))
-import           ZkFold.Base.Protocol.Protostar.FiatShamir             (FiatShamir (..))
-import           ZkFold.Base.Protocol.Protostar.IVC                    (IVCResult, ivcVerify)
-import           ZkFold.Prelude                                        (replicate)
+import           ZkFold.Base.Algebra.Basic.Number                (KnownNat, type (+), type (-), value)
+import           ZkFold.Base.Data.ByteString                     (Binary)
+import           ZkFold.Base.Data.Vector                         (Vector, unsafeToVector)
+import           ZkFold.Base.Protocol.IVC.ArithmetizableFunction (ArithmetizableFunction (ArithmetizableFunction))
+import           ZkFold.Base.Protocol.IVC.Commit                 (HomomorphicCommit)
+import           ZkFold.Base.Protocol.IVC.CommitOpen             (CommitOpen (..))
+import           ZkFold.Base.Protocol.IVC.FiatShamir             (FiatShamir (..))
+import           ZkFold.Base.Protocol.IVC.Internal               (IVCResult, ivcVerify)
+import           ZkFold.Prelude                                  (replicate)
 import           ZkFold.Symbolic.Class
 import           ZkFold.Symbolic.Compiler
-import           ZkFold.Symbolic.Data.Class                            (SymbolicData (..))
-import           ZkFold.Symbolic.Data.FieldElement                     (FieldElement (..))
-import           ZkFold.Symbolic.Data.Input                            (SymbolicInput)
-import           ZkFold.Symbolic.Interpreter                           (Interpreter (..))
+import           ZkFold.Symbolic.Data.Class                      (SymbolicData (..))
+import           ZkFold.Symbolic.Data.FieldElement               (FieldElement (..))
+import           ZkFold.Symbolic.Data.Input                      (SymbolicInput)
+import           ZkFold.Symbolic.Interpreter                     (Interpreter (..))
 
 -- | Takes a function `f` and returns a circuit `C` with input `y` and witness `w`.
 -- The circuit is such that `C(y, w) = 0` implies that `y = x(n)` for some positive `n` where
