+++ symbolic-base/src/ZkFold/Base/Protocol/IVC/SpecialSound.hs
@@ -2,18 +2,18 @@
 {-# LANGUAGE TypeOperators        #-}
 {-# LANGUAGE UndecidableInstances #-}
 
-module ZkFold.Base.Protocol.Protostar.SpecialSound where
+module ZkFold.Base.Protocol.IVC.SpecialSound where
 
-import           Data.Functor.Rep                                      (Representable (..))
-import           Data.Map.Strict                                       (elems)
-import           GHC.Generics                                          ((:*:) (..))
-import           Prelude                                               (($))
+import           Data.Functor.Rep                                (Representable (..))
+import           Data.Map.Strict                                 (elems)
+import           GHC.Generics                                    ((:*:) (..))
+import           Prelude                                         (($))
 
 import           ZkFold.Base.Algebra.Basic.Class
 import           ZkFold.Base.Algebra.Basic.Number
-import           ZkFold.Base.Data.Vector                               (Vector)
-import qualified ZkFold.Base.Protocol.Protostar.AlgebraicMap           as AM
-import           ZkFold.Base.Protocol.Protostar.ArithmetizableFunction (ArithmetizableFunction (..))
+import           ZkFold.Base.Data.Vector                         (Vector)
+import qualified ZkFold.Base.Protocol.IVC.AlgebraicMap           as AM
+import           ZkFold.Base.Protocol.IVC.ArithmetizableFunction (ArithmetizableFunction (..))
 import           ZkFold.Symbolic.Class
 import           ZkFold.Symbolic.Compiler
