--- symbolic-base/src/ZkFold/Base/Protocol/Protostar.hs
@@ -1,3 +1,0 @@
-module ZkFold.Base.Protocol.Protostar (module P) where
-
-import           ZkFold.Base.Protocol.Protostar.IVC as P
