+++ symbolic-base/symbolic-base.cabal
@@ -119,6 +119,19 @@ library
       ZkFold.Base.Data.Type
       ZkFold.Base.Data.Utils
       ZkFold.Base.Data.Vector
+      ZkFold.Base.Protocol.IVC
+      ZkFold.Base.Protocol.IVC.Accumulator
+      ZkFold.Base.Protocol.IVC.AccumulatorScheme
+      ZkFold.Base.Protocol.IVC.AlgebraicMap
+      ZkFold.Base.Protocol.IVC.ArithmetizableFunction
+      ZkFold.Base.Protocol.IVC.CommitOpen
+      ZkFold.Base.Protocol.IVC.Commit
+      ZkFold.Base.Protocol.IVC.FiatShamir
+      ZkFold.Base.Protocol.IVC.Internal
+      ZkFold.Base.Protocol.IVC.NARK
+      ZkFold.Base.Protocol.IVC.Oracle
+      ZkFold.Base.Protocol.IVC.RecursiveCircuit
+      ZkFold.Base.Protocol.IVC.SpecialSound
       ZkFold.Base.Protocol.KZG
       ZkFold.Base.Protocol.Plonk
       ZkFold.Base.Protocol.Plonk.Prover
@@ -142,19 +155,6 @@ library
       ZkFold.Base.Protocol.Plonkup.Verifier.Commitments
       ZkFold.Base.Protocol.Plonkup.Verifier.Setup
       ZkFold.Base.Protocol.Plonkup.Witness
-      ZkFold.Base.Protocol.Protostar
-      ZkFold.Base.Protocol.Protostar.Accumulator
-      ZkFold.Base.Protocol.Protostar.AccumulatorScheme
-      ZkFold.Base.Protocol.Protostar.AlgebraicMap
-      ZkFold.Base.Protocol.Protostar.ArithmetizableFunction
-      ZkFold.Base.Protocol.Protostar.CommitOpen
-      ZkFold.Base.Protocol.Protostar.Commit
-      ZkFold.Base.Protocol.Protostar.FiatShamir
-      ZkFold.Base.Protocol.Protostar.IVC
-      ZkFold.Base.Protocol.Protostar.NARK
-      ZkFold.Base.Protocol.Protostar.Oracle
-      ZkFold.Base.Protocol.Protostar.RecursiveCircuit
-      ZkFold.Base.Protocol.Protostar.SpecialSound
       ZkFold.Base.Protocol.NonInteractiveProof
       ZkFold.Base.Protocol.NonInteractiveProof.Internal
       ZkFold.Base.Protocol.NonInteractiveProof.Prover
@@ -250,11 +250,11 @@ test-suite symbolic-base-test
       Tests.Field
       Tests.GroebnerBasis
       Tests.Group
+      Tests.IVC
       Tests.NonInteractiveProof
       Tests.Pairing
       Tests.Permutations
       Tests.Plonkup
-      Tests.Protostar
       Tests.RSA
       Tests.SHA2
       Tests.UInt
